requirements.txt
msal==1.25.0
requests==2.31.0
pandas==2.1.4
openpyxl==3.1.2
jinja2==3.1.2
python-dotenv==1.0.0 